+++ tests/test_scheduler_runtime.py
@@ -20,6 +20,7 @@ from .scheduler_runtime_progress_recovery import SchedulerRuntimeTestsPart10
 from .scheduler_runtime_rebuild_cancellation import SchedulerRuntimeTestsPart11
 from .scheduler_runtime_command_workers import SchedulerRuntimeTestsPart12
 from .scheduler_runtime_manual_compatibility import SchedulerRuntimeTestsPart13
+from .scheduler_runtime_heartbeat_continuity import SchedulerRuntimeTestsPart14
 from .scheduler_runtime_removed_coupling_tests import RemovedCouplingTests
 
 
@@ -38,6 +39,7 @@ class SchedulerRuntimeTests(
     SchedulerRuntimeTestsPart11,
     SchedulerRuntimeTestsPart12,
     SchedulerRuntimeTestsPart13,
+    SchedulerRuntimeTestsPart14,
     unittest.TestCase,
 ):
     pass